using csv to write result.
